Added UML Diagram
I've started having to plot out the project in Lucidchart.  Hopefully this will lead to a better structure than my current (chaotic) approach.  There's also a few odds and ends with refactoring of the code, but there will probably be a very major re-write of most of the code soon anyway.