Started Adding "Pixels" Rendering Library
Got the graphics side of things to the point where an empty window can be displayed for a brief period of time (about a second) before it crashes.  I still can't render anything to it, and I'm giving more thought to exactly what the project is supposed to involve and what should, and should not, be imported into it.