Can Now Display Empty Window
I've got winit working (to the point of displaying an empty window that doesn't immediately crash), although it heavily relies on Rust syntax I'm not familiar with (due to it being lifted out of the online examples).  The aim is now to make it code that I'm more comfortable manipulating.